Fix PR#1519 for increase swift device size (#480)
This patch fixes the issue mentioned in PR#1519.

The default size of devstack swift is 2G, if we enable octavia or we
create some new images into glance during test. We will get failure.
The key reason for this is the swift device size is too small. So this
patch will increase the size through pass a env parameter in devstack
local conf.
Close: theopenlab/openlab#222