Cleanup Bazel related roles (#496)
Bazel is now packed in base images for CNCF project tests now, and the tests are successfully running, we can get rid of the related role for installing it to make the repo clean and easier to maintain. If the current way of doing the tests is not good enough, we can come up with other ways like build bazel in disk-image-builder or try to get an official bazel source in Ubuntu deb to make the whole process simpler.