Fix openlab-check job (#1120)
The upstream ansible-lint has been upgraded to 5.0.1 version. And request the min version of ansible is 2.9.0 .
Openlab ansible runtime is 2.8.0 version, and all the jobs code with 2.8.0 syntax and format, so for ansible-lint, we can not upgrade to the latest version(5.0.1) directly.

And from my test, there are several obvious bugs towards the latest version, such as the `skip_list` doesn't work (expect is exchanging by ErrId), `tag_list` can only use the said newer names, and the initial check error in https://github.com/ansible-community/ansible-lint/blob/master/src/ansiblelint/_prerun.py#L39 (If we had a installed ansible version, the code would be failure.) . So I think the upstream software quality of ansible-lint is unavailable. We need to wait until the software is stable before upgrading the ansible version.

So current solution is keeping to use the old version(4.3.7), which is latest version for 4.X . The differences between ansible-lint 5.0.1 and ansible 4.3.7 are very big, the same as ansible 2.9 and ansible 2.8 .