Add a workaround to avoid issues when devstack needs APT
It's rare but it can happen that devstack runs at the same time as a
crontab is triggered and runs APT to e.g. execute unattended upgrades of
the system.

Devstack would fail since APT isn't available.
This patch adds a workaround so it'll remove the crontabs before
deploying devstack.